특정 버전 완전 삭제 기능 추가

+++ install.sh
@@ -20,8 +20,24 @@ sudo chown $CURRENT_USER.$CURRENT_GROUP $JDK_SH_ROOT
 
 # 기존 스크립트 및 version.properties 파일 제거
 rm -f $JDK_SH_ROOT/update.sh $JDK_SH_ROOT/clean.sh $JDK_SH_ROOT/version.properties
-wget https://raw.githubusercontent.com/fp024/simple-jdk-manager/master/update.sh -O $JDK_SH_ROOT/update.sh || { echo "[오류] update.sh 파일 다운로드에 실패했습니다."; exit 1; }
-wget https://raw.githubusercontent.com/fp024/simple-jdk-manager/master/clean.sh -O $JDK_SH_ROOT/clean.sh || { echo "[오류] clean.sh 파일 다운로드에 실패했습니다."; exit 1; }
+
+# update.sh 복사 또는 다운로드
+if [ -f "./update.sh" ]; then
+  echo "[알림] 현재 디렉토리의 update.sh를 복사하여 설치합니다."
+  cp ./update.sh $JDK_SH_ROOT/update.sh
+else
+  echo "[알림] update.sh를 다운로드하여 설치합니다."
+  wget https://raw.githubusercontent.com/fp024/simple-jdk-manager/master/update.sh -O $JDK_SH_ROOT/update.sh || { echo "[오류] update.sh 파일 다운로드에 실패했습니다."; exit 1; }
+fi
+
+# clean.sh 복사 또는 다운로드
+if [ -f "./clean.sh" ]; then
+  echo "[알림] 현재 디렉토리의 clean.sh를 복사하여 설치합니다."
+  cp ./clean.sh $JDK_SH_ROOT/clean.sh
+else
+  echo "[알림] clean.sh를 다운로드하여 설치합니다."
+  wget https://raw.githubusercontent.com/fp024/simple-jdk-manager/master/clean.sh -O $JDK_SH_ROOT/clean.sh || { echo "[오류] clean.sh 파일 다운로드에 실패했습니다."; exit 1; }
+fi
 
 chmod u+x $JDK_SH_ROOT/update.sh $JDK_SH_ROOT/clean.sh
